[ResponseOps][Connectors] Fix accessTokenUrl validation (#258290)
## 📄 Summary

Fix Webhook Connector `accessTokenUrl` validation

<details>
<summary>

## 🧪 Verification steps

</summary>

### ✅ Happy Path

- Create or update a Webhook connector with OAuth2 client credentials
where both `url` and `accessTokenUrl` use hosts listed in
`xpack.actions.allowedHosts`. Validation passes and the connector is
saved.

### ⚡️ Edge Cases

- Webhook with OAuth2 and only main URL in allowedHosts: if
`accessTokenUrl` host is not allowed, creation/update fails with a clear
error.
- Webhook without OAuth2: behavior unchanged; only main URL is validated
against allowedHosts.

### ❌ Failure Cases

- Create a Webhook connector with OAuth2 and an `accessTokenUrl` whose
host is not in `xpack.actions.allowedHosts`. Expect 400 with message
containing the token URL and "is not added to the Kibana config" (or
similar wording from the actions plugin).

</details>

## ⏪ Backport rationale

Backporting to the track where the issue appeared

## Release Notes

Fix Webhook Connector `accessTokenUrl` validation

## ☑️ Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [ ] [Flaky Test
Runner](https://ci-stats.kibana.dev/trigger_flaky_test_runner/1) was
used on any tests changed
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)
- [x] Review the [backport
guidelines](https://docs.google.com/document/d/1VyN5k91e5OVumlc0Gb9RPa3h1ewuPE705nRtioPiTvY/edit?usp=sharing)
and apply applicable `backport:*` labels.

(cherry picked from commit cbbdd5317c9181efc6c3d2b78412fb37d4d4eb51)

# Conflicts:
#	x-pack/platform/plugins/shared/stack_connectors/server/connector_types/webhook/validations.ts
