+++ x-pack/platform/test/alerting_api_integration/security_and_spaces/group2/tests/actions/connector_types/webhook.ts
@@ -329,6 +329,32 @@ export default function webhookTest({ getService }: FtrProviderContext) {
       expect(result.message).to.match(/is not added to the Kibana config/);
     });
 
+    it('should reject OAuth2 webhook when accessTokenUrl host is not in allowedHosts', async () => {
+      const tokenUrl = 'http://oauth.mynonexistent.com/token';
+      const { body: result } = await supertest
+        .post('/api/actions/connector')
+        .set('kbn-xsrf', 'test')
+        .send({
+          name: 'OAuth2 Webhook disallowed token URL',
+          connector_type_id: '.webhook',
+          secrets: {
+            clientSecret: 'secret',
+          },
+          config: {
+            url: webhookSimulatorURL,
+            hasAuth: true,
+            authType: 'webhook-oauth2-client-credentials',
+            accessTokenUrl: tokenUrl,
+            clientId: 'client-id',
+          },
+        })
+        .expect(400);
+
+      expect(result.error).to.eql('Bad Request');
+      expect(result.message).to.contain(tokenUrl);
+      expect(result.message).to.match(/is not added to the Kibana config/);
+    });
+
     it('should handle unreachable webhook targets', async () => {
       const webhookActionId = await createWebhookAction(
         'http://some.non.existent.com/endpoint',
